+++ src/primitives/sector/points.rs
@@ -7,7 +7,7 @@ use crate::{
 };
 
 /// Iterator over all points inside the sector.
-#[derive(Copy, Clone, PartialEq, PartialOrd, Debug)]
+#[derive(Clone, PartialEq, Debug)]
 pub struct Points {
     iter: DistanceIterator,