+++ src/primitives/sector/styled.rs
@@ -14,7 +14,7 @@ use crate::{
 };
 
 /// Pixel iterator for each pixel in the sector border
-#[derive(Copy, Clone, PartialEq, PartialOrd, Debug)]
+#[derive(Clone, PartialEq, Debug)]
 pub struct StyledPixels<C>
 where
     C: PixelColor,